move and clean up tests
